Moved garbage collection-related methods into a trait
Object subtypes which are garbage collected now implement the trait
`GarbageCollected`, which includes all the methods for marking and
checking.
Note that this trait does not include a method `dealloc` (or similar),
which is a method of the `Object` type and will remain so, as it uses
pointers to structs instead of the structs themselves.