RlispError type ; rust_macros module
Big changes! We now have `RlispError` as a variant of `Object`, and
errors during evaluation result in those instead of returning
`Result<Object>`s. This allows runtime error-handling in a try-catch
style in the future, at the cost of moving away from Rust's beautiful
strong-type error system and into having to manually check for errors.